requirements.txt
google-genai
Pillow
elevenlabs
python-dotenv
